+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/php/webhook.php
@@ -1,7 +1,0 @@
-<?php
-
-$container->loadFromExtension('framework', [
-    'webhook' => ['enabled' => true],
-    'http_client' => ['enabled' => true],
-    'serializer' => ['enabled' => true],
-]);