+++ src/Symfony/Component/Webhook/Tests/Fixtures/TestRequestParser.php
@@ -1,0 +1,36 @@
+<?php
+
+/*
+ * This file is part of the Symfony package.
+ *
+ * (c) Fabien Potencier <fabien@symfony.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Symfony\Component\Webhook\Tests\Fixtures;
+
+use Symfony\Component\HttpFoundation\ChainRequestMatcher;
+use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpFoundation\RequestMatcherInterface;
+use Symfony\Component\RemoteEvent\RemoteEvent;
+use Symfony\Component\Webhook\Client\AbstractRequestParser;
+use Symfony\Component\Webhook\Exception\RejectWebhookException;
+
+class TestRequestParser extends AbstractRequestParser
+{
+    protected function getRequestMatcher(): RequestMatcherInterface
+    {
+        return new ChainRequestMatcher([]);
+    }
+
+    protected function doParse(Request $request, #[\SensitiveParameter] string $secret): ?RemoteEvent
+    {
+        if ($secret !== $request->headers->get('X-Secret')) {
+            throw new RejectWebhookException(406, 'Invalid secret.');
+        }
+
+        return new RemoteEvent('parsed', 'the-id', json_decode($request->getContent(), true));
+    }
+}